+++ datafusion/core/tests/physical_optimizer/mod.rs
@@ -16,3 +16,4 @@
 // under the License.
 
 mod aggregate_statistics;
+mod limit_pushdown;
